Code adjustment, added package, updated folder path.
Moved return of "oops..." text from search button to results window, added package declaration, updated folder path to match package name.
